ENH: Add option to hide the maximize view button
In custom apps usually it is easy to simply find unwanted toolbar elements by object name and hide it. However, in the case of the maximize view button, it is automatically shown if the view can be maximized, which makes it reappear in the custom apps. It is possible to remove it from the layout but it does not work properly in all cases. By adding such an option it is possible to hide it appropriately.